# docs/swagger.yaml
definitions:
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /config/keys/get:
    post:
      description: get config from backend
      parameters:
      - description: settings
        in: body
        name: settings
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: get config from backend
      tags:
      - config
  /config/keys/set:
    post:
      description: set config to backend
      parameters:
      - description: settings
        in: body
        name: settings
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: set config to backend
      tags:
      - config
...
swagger: "2.0"